tests fixed a bunch

+++ maidenhead.h
@@ -3,14 +3,14 @@
 
 
 typedef struct latlon {
-	float lat;
-	float lon;
+	double lat;
+	double lon;
 } latlon;
 
 
 // parse a nul terminated string like "FN41sp" to the appropriate lat-lon.
 latlon maidenhead_to_latlon( char * loc );
-void maidenhead_to_lat_lon( char * loc, float * lat, float * lon );
+void maidenhead_to_lat_lon( char * loc, double * lat, double * lon );
 
 // encode a lat-lon into a maidenhead grid square
 // precision is how many levels to encode, e.g. FN41 is 2, FN41sp is 3, etc
@@ -20,10 +20,10 @@ void maidenhead_to_lat_lon( char * loc, float * lat, float * lon );
 // TODO?: latlon_to_maidenhead_locator doesn't set that nul!
 // initialize the buffer to 0, or keep track of the string length yourself to avoid issues
 void latlon_to_maidenhead( latlon in, char * maidenhead_out, int precision );
-void lat_lon_to_maidenhead( float lat, float lon, char * maidenhead_out, int precision );
+void lat_lon_to_maidenhead( double lat, double lon, char * maidenhead_out, int precision );
 
 //
-float distance_between_maidenhead_locators_in_subsquares(char*a,char*b);
+double distance_between_maidenhead_locators_in_subsquares(char*a,char*b);
 int maidenhead_locators_are_adjacent( char *a, char *b);
 
 #endif
